+++ source/lang/src/frontend/node/expression.c
@@ -3,13 +3,13 @@
 
 #include "node.h"
 
-Expression* newExpresion( NodeType kind ) {
+Expression* newExpression( NodeType kind ) {
     Expression* out = (Expression*)malloc(sizeof(Expression));
     if (out==NULL) {
         fprintf(stderr, "Could not allocate memory for new Expression");
         return NULL;
     }
-    out->exprKind->kind = kind;
+    out->exprKind = newStatement(kind);
     return out;
 }
