+++ source/lang/src/frontend/token/deleteHead.c
