+++ source/lang/src/frontend/token/deleteTail.c
