+++ source/lang/src/lang.h
@@ -1,6 +1,6 @@
 #ifndef _lang_h
 #define _lang_h
 
-#include "frontend/frontend.h"
+#include "frontend/node/node.h"
 
 #endif
